Reviewed by Ken.
        - fixed 3167025 -- REGRESSION: http://www.buypcdirect.com doesn't wrap correctly b/c of huge cell width value

        * khtml/rendering/table_layout.h: Make the fields in the Layout structure used in table
        width computations use int instead of short.
        * khtml/rendering/table_layout.cpp:
        (FixedTableLayout::calcMinMaxWidth): Make sure we don't overflow m_minWidth or m_maxWidth in case the
        specified fixed width value is huge.
        (FixedTableLayout::layout): Use ints for the calcWidth array instead of shorts.
        (AutoTableLayout::calcMinMaxWidth): Clamp both maxWidth and minWidth to 0x7fff to match the other code in
        this file, rather than using 10000.


Canonical link: https://commits.webkit.org/3345@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@3593 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/khtml/rendering/table_layout.h
@@ -94,11 +94,11 @@ protected:
 		   calcWidth( 0 ) {}
 	Length width;
 	Length effWidth;
-	short minWidth;
-	short maxWidth;
-	short effMinWidth;
-	short effMaxWidth;
-	short calcWidth;
+	int minWidth;
+	int maxWidth;
+	int effMinWidth;
+	int effMaxWidth;
+	int calcWidth;
     };
 
     QMemArray<Layout> layoutStruct;
